fft screwed up

+++ grab.h
@@ -4,7 +4,7 @@
 struct audio_stream {
   unsigned int buffer_length;
   unsigned int sample_rate;
-  float *audio_data;
+  double *audio_data;
 };
 int mic_setup ( void (*dpf)(struct audio_stream*) );
 void mic_grab (void);
